manually handle frameskipping for better audio on non multiple of 60hz screens
