+++ website/app/src/mgba.tsx
@@ -5,10 +5,10 @@ import {
   useRef,
   useState,
 } from "react";
-import mGBA from "./vendor/mgba";
+import mGBA, { mGBAEmulator } from "./vendor/mgba";
 import { GbaKey, KeyBindings } from "./bindings";
 import { styled } from "styled-components";
-import { useSmoothedFramerate } from "./useSmoothedFramerate.hook";
+import { useFrameSkip } from "./useFrameSkip.hook";
 
 type Module = any;
 
@@ -41,18 +41,10 @@ export interface MgbaHandle {
   buttonRelease: (key: GbaKey) => void;
 }
 
-const whichFrameSkip = (frameRate: number): number | undefined => {
-  if ((frameRate + 5) % 60 <= 10) {
-    // framerate close to multiple of 60
-    // use frameskip
-    return Math.round(frameRate / 60);
-  }
-};
-
 export const Mgba = forwardRef<MgbaHandle, MgbaProps>(
   ({ gameUrl, volume, controls, paused }, ref) => {
     const canvas = useRef(null);
-    const mgbaModule = useRef<Module>({});
+    const mgbaModule = useRef<Module>({} as mGBAEmulator);
 
     const [state, setState] = useState(MgbaState.Uninitialised);
     const [gameLoaded, setGameLoaded] = useState(false);
@@ -101,23 +93,7 @@ export const Mgba = forwardRef<MgbaHandle, MgbaProps>(
         };
     }, [state]);
 
-    const frameRate = useSmoothedFramerate();
-    const frameSkipToUse = whichFrameSkip(frameRate);
-
-    useEffect(() => {
-      if (!gameLoaded) return;
-
-      if (frameSkipToUse) {
-        // framerate close to multiple of 60
-        // use frameskip
-        console.log("Using frameskip");
-        mgbaModule.current.setMainLoopTiming(1, frameSkipToUse);
-      } else {
-        // frame rate not close to multiple of 60, use timeout
-        console.log("Using timeout");
-        mgbaModule.current.setMainLoopTiming(0, 1000 / 59.727500569606);
-      }
-    }, [frameSkipToUse, gameLoaded]);
+    useFrameSkip(mgbaModule);
 
     useEffect(() => {
       if (!gameLoaded) return;
